+++ src/constants/links.ts
@@ -27,13 +27,13 @@ export const LINKS = {
   ICE_FAQ: 'https://ice.io/faq',
   STATUS_NOTICE: 'https://ice-production.b-cdn.net/assets/notice.json',
   BETA_TESTING: 'https://ice.io/beta-testing',
-  FACEBOOK_PAGE: `fb://profile/${ENV.FACEBOOK_PAGE_ID}`,
-  FACEBOOK_WEB_PAGE: 'https://facebook.com/ice.blockchain',
-  INSTAGRAM_PAGE: 'instagram://user?username=ice.blockchain',
-  INSTAGRAM_WEB_PAGE: 'https://www.instagram.com/ice.blockchain',
-  LINKEDIN_PAGE: 'linkedin://company/ice-blockchain',
-  LINKEDIN_WEB_PAGE: 'https://linkedin.com/company/ice-blockchain',
-  YOUTUBE_PAGE: `vnd.youtube://channel/${ENV.YOUTUBE_CHANNEL_ID}`,
-  YOUTUBE_WEB_PAGE: 'https://youtube.com/@ice.blockchain',
-  TIKTOK_PAGE: 'https://www.tiktok.com/@ice.blockchain',
+  FACEBOOK_APP: `fb://profile/${ENV.FACEBOOK_PAGE_ID}`,
+  FACEBOOK_WEB: 'https://facebook.com/ice.blockchain',
+  INSTAGRAM_APP: 'instagram://user?username=ice.blockchain',
+  INSTAGRAM_WEB: 'https://www.instagram.com/ice.blockchain',
+  LINKEDIN_APP: 'linkedin://company/ice-blockchain',
+  LINKEDIN_WEB: 'https://linkedin.com/company/ice-blockchain',
+  YOUTUBE_APP: `vnd.youtube://channel/${ENV.YOUTUBE_CHANNEL_ID}`,
+  YOUTUBE_WEB: 'https://youtube.com/@ice.blockchain',
+  TIKTOK_WEB: 'https://www.tiktok.com/@ice.blockchain',
 };
