+++ src/screens/Modals/PopUp/index.tsx
@@ -45,7 +45,6 @@ export type PopUpProps = {
   showCheckbox?: boolean;
   checkboxText?: string;
   onDismiss?: () => void;
-  onClose?: () => void;
 };
 
 export const PopUp = () => {
@@ -66,7 +65,6 @@ export const PopUp = () => {
       showCheckbox = false,
       checkboxText = '',
       onDismiss,
-      onClose,
     },
   } = useRoute<RouteProp<MainStackParamList, 'PopUp'>>();
   const navigation = useNavigation();
@@ -156,7 +154,7 @@ export const PopUp = () => {
             <CloseButton
               style={styles.closeButton}
               onPress={() => {
-                onClose?.();
+                onDismiss?.();
               }}
             />
           )}
