+++ src/store/modules/Socials/data.tsx
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: ice License 1.0
 
+import {LINKS} from '@constants/links';
 import {Images} from '@images';
 import {SocialType} from '@store/modules/Socials/types';
 import {t} from '@translations/i18n';
@@ -10,6 +11,8 @@ export interface SocialData {
   title: string;
   description: string;
   buttonTitle: string;
+  linkApp: string;
+  linkWeb: string;
 }
 const facebookDescription = () => {
   return `${t('social_media.facebook.description_part1')}\n\n${t(
@@ -29,29 +32,39 @@ export const socialData: Record<SocialType, SocialData> = {
     title: t('social_media.tiktok.title'),
     description: t('social_media.tiktok.description_part1'),
     buttonTitle: t('button.follow_us'),
+    linkApp: LINKS.TIKTOK_WEB,
+    linkWeb: LINKS.TIKTOK_WEB,
   },
   youtube: {
     image: Images.social.youtube,
     title: t('social_media.youtube.title'),
     description: t('social_media.youtube.description_part1'),
     buttonTitle: t('button.subscribe'),
+    linkApp: LINKS.YOUTUBE_APP,
+    linkWeb: LINKS.YOUTUBE_WEB,
   },
   linkedin: {
     image: Images.social.linkedin,
     title: t('social_media.linkedin.title'),
     description: t('social_media.linkedin.description_part1'),
     buttonTitle: t('button.follow_us'),
+    linkApp: LINKS.LINKEDIN_APP,
+    linkWeb: LINKS.LINKEDIN_WEB,
   },
   facebook: {
     image: Images.social.facebook,
     title: t('social_media.facebook.title'),
     description: facebookDescription(),
     buttonTitle: t('button.follow_us'),
+    linkApp: LINKS.FACEBOOK_APP,
+    linkWeb: LINKS.FACEBOOK_WEB,
   },
   instagram: {
     image: Images.social.instagram,
     title: t('social_media.instagram.title'),
     description: instagramDescription(),
     buttonTitle: t('button.follow_us'),
+    linkApp: LINKS.INSTAGRAM_APP,
+    linkWeb: LINKS.INSTAGRAM_WEB,
   },
 };
